change to one feature withoutexactarithmetic
