+++ src/math/log_div.rs
@@ -1,8 +1,5 @@
-#[cfg(not(any(feature = "exact", feature = "approximate")))]
-compile_error!("At least one of the features \"exact\" and \"approximate\" must be enabled.");
-
-#[cfg(all(feature = "exact", feature = "approximate"))]
+#[cfg(not(feature = "withoutexactarithmetic"))]
 pub type LogDiv = super::log_div_enum::LogDivEnum;
 
-#[cfg(all(not(feature = "exact"), feature = "approximate"))]
+#[cfg(feature = "withoutexactarithmetic")]
 pub type LogDiv = super::log_div_f64::LogDivF64;
